2001-12-16  Alexander R Angas  <wgd@internode.on.net>
	* src/sound/event_ss.c:
	* src/sound/event_ss_win32.c: Added sound save and restore.

	* src/sound/sfx_save.c:
	* src/sound/sfx_save.cfsml: New save game version that works on both
	types of sound servers. Still reads old saved games. No crashes yet.
	Problems when restoring on occasions.

	* src/sound/sound.c: Added song_lib_count().

	* src/sound/soundserver.c: Clean ups, rearrangements, minor bug fixes.

	* src/include/soundserver.h: Fixed some variable types and added some
	function declarations.

	* src/sound/polled_ss.c: Now uses functions defined in soundserver.c
	for most sound commands.

	* src/sound/polled_ss_win32.c: Minor synchronisation fixes.

	* src/sound/oldmidi.c (map_MIDI_instruments): Commented out creation of
	freesci_instr log file which does nothing.

	* src/scicore/tools.c (sci_find_first): Commented out ENOENT error
	except on debug builds.

	* src/include/resource.h: Fixed type conversion warning.

	* src/config.c:
	* src/config.l: Added SATISFY_PURIFY to fix UMC.

	* src/win32/fsci_dll.dsp:
	* src/win32/sciv.dsp: Added 'Purify' configuration. Removed HAVE_DDRAW
	from Release builds.

	* src/scicore/sci_memory.c: Fixed error that occurred with new Purify
	configuration.

	* ./README.Win32: Added info on above configuration and sound problems.

+++ src/include/resource.h
@@ -191,7 +191,7 @@ typedef guint16 word;
 static inline gint16
 getInt16(byte *d)
 {
-  return *d | (d[1] << 8);
+	return (short)(*d | (d[1] << 8));
 }
 
 #define getUInt16(d) (guint16)(getInt16(d))
